Move test constants to a dedicated package
